+++ protected/humhub/components/console/UrlManager.php
@@ -10,6 +10,7 @@ namespace humhub\components\console;
 
 use humhub\components\InstallationState;
 use Yii;
+use yii\base\InvalidConfigException;
 
 /**
  * @inheritdoc
@@ -21,6 +22,10 @@ class UrlManager extends \humhub\components\UrlManager
      */
     public function init()
     {
+        if (Yii::$app instanceof \yii\web\Application) {
+            throw new InvalidConfigException('Console UrlManager cannot be used in a web application');
+        }
+
         $urlParts = parse_url($this->getConfiguredBaseUrl());
 
         $this->setBaseUrl($urlParts['path'] ?? '');
@@ -30,7 +35,7 @@ class UrlManager extends \humhub\components\UrlManager
             $hostInfo .= ':' . $urlParts['port'];
         }
         $this->setHostInfo($hostInfo);
-        $this->setScriptUrl($this->getBaseUrl() . (str_replace($this->getBaseUrl(), '', $this->getScriptUrl()) ?: '/index.php'));
+        $this->setScriptUrl($this->getBaseUrl() . ($this->getScriptUrl() ?: '/index.php'));
 
         parent::init();
     }